sample-app/src/order-service/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
python-json-logger
